persist: move PersistParameters into persist
Having this inside persist will let us hook things up with a much
smaller increase to our public API surface area.
